cyber/requirements.txt
pytest~=7.4.3
Pillow~=10.1.0